Added otp verification

+++ models/User.js
@@ -52,6 +52,15 @@ module.exports = class Users {
   static findOne(emailId) {
     return db.execute("SELECT * FROM user_ WHERE email = ?", [emailId]);
   }
+  static findByEmail(email){
+    return db.query(`
+      SELECT u.*, r.name AS role
+      FROM user_ u
+      LEFT JOIN user_roles r ON u.role_id = r.id
+      WHERE u.email = ?
+    `, [email]);
+    
+    }
 
   static findAll() {
     return db.execute("SELECT * FROM user_");
@@ -106,27 +115,33 @@ module.exports = class Users {
   
 
   static async verifyOtp(emailId, otp) {
-    const response = await db.execute("SELECT * FROM `cloud_enviro_otp` WHERE email = ?", [emailId]);
-    const { expires_at, otp: existingOtp } = response[0][0];
+    const [rows] = await db.execute("SELECT * FROM `cloud_enviro_otp` WHERE email = ?", [emailId]);
+  
+    if (!rows || rows.length === 0) {
+      return { success: false, message: "OTP expired" };
+    }
+  
+    const { expires_at, otp: existingOtp } = rows[0];
+  
     const expiringTime = new Date(expires_at).getTime();
     const currentTime = Date.now();
+  
     if (currentTime > expiringTime) {
-      // Delete data in cloud_enviro_otp if it otp expires !
-      await db.execute("DELETE FROM `cloud_enviro_otp` WHERE `email` = ?;", [emailId]);
+      await db.execute("DELETE FROM `cloud_enviro_otp` WHERE `email` = ?", [emailId]);
       return { success: false, message: "OTP has expired." };
     }
-
+  
     if (otp.toString() !== existingOtp.toString()) {
       return { success: false, message: "Invalid OTP." };
     }
-
-    // Delete email and otp if it verifies successfully!
-    await db.execute("DELETE FROM `cloud_enviro_otp` WHERE `email` = ?;", [emailId]);
+  
+    await db.execute("DELETE FROM `cloud_enviro_otp` WHERE `email` = ?", [emailId]);
     return { success: true, message: "OTP verified successfully." };
   }
+  
 
   static async generateOtp(emailId) {
-    const otp = randomInt().toString();
+    const otp = randomInt(0, 10000).toString().padStart(4, '0');
     const expiresAtUnix = Date.now() + 10 * 60 * 1000;
     const dateTime = new Date(expiresAtUnix);
     const expiresAt =
@@ -162,4 +177,5 @@ module.exports = class Users {
       return { success: true, msg: "Created", otp: otp };
     }
   }
+  
 };
